lisp-interaction-mode: indent using spaces
Many users set `indent-tabs-mode' global value to nil.  And I agree with
those users, this should have been the default.  But the reality is that
it isn't the default.

If you change the default locally, then your default is different from
the "default default", which is a problem because most files that were
authored by people who prefer to use tabs do NOT explicitly set the
file local value to t, because that doing so would (theoretically) be
redundant.

If you do set `indent-tabs-mode' to nil despite this, then you will end
up using spaces for indentation in files whose author decided to use
tabs.  What I do instead is to set the file-/buffer-local value for all
files whose author I am.  That's a bit inconvenient, but if I didn't do
that then I would have to check every contributions for tabs.