git-commit: no longer set value of two variables
The function on `git-commit-finish-query-functions' by default no longer
complains about "too long" summary lines (which is why we set this to
nil before) but it still complains about non-empty second line (which is
a good thing).

The default value of `git-commit-summary-max-length' has been increased
from 50 to 68, which is close enough to the 64 we used here before.