epkg: no longer set epkg-repository explicitly
Updated `no-littering' puts that repository in "var/" already, so
we do not have to do that anymore.  However a user (i.e. me) might
want to put that repository yet somewhere else and to enable that
`no-littering' has to be loaded before `epkg' is configured, so
switch the positions of the two `use-package' forms.