Emacs.g: Move Makefile and borg-bootstrap into borg package
In the borg repository these files are renamed to "borg.mk" and
"borg.sh" and modified slightly.  Add a new "Makefile" here.  It
defines one simple target `borg-bootstrap', which is required to
make "borg.mk" and "borg.sh" available during bootstrap and sources
"borg.mk" for all the other targets it used to define.  The target
defined here is fairly simple and hopefully won't have to be updated
often, but independent configurations still have integrate a copy of
this file, because we cannot bootstrap out of thin air and `git
submodule' isn't up to the task yet.

Update borg to v2.0.0-20-g8399ec8