Update undo-tree to release/0.6.6-7-gcc7354c0
Because the two official upstream repositories (the author's and the
one on GNU Elpa) are never in sync I am now treating my own fork on
the Emacsorphanage as the upstream.

That fork also has a few minor patches, which I could not get the
author to merge (or just comment on).  They are not very important
except that without them undo-tree would be the only package that
I have installed that cannot be compiled without warnings.  After
several years of hoping the author would do something about that
eventually I am giving up.

I intend to merge changes from the upstream repositories, but I might
not instantly notice when new commits appear.  You might have to give
me a nudge (to which I will reply at the very least).