smerge-mode: Do not extend smerge-redefine-changed
By default this face is not neither used nor has a visible effect.
It is used if it does `face-differs-from-default-p', in which case
it replaces `smerge-refined-removed' and `smerge-refined-added'.

Setting `:extend' makes the `face-differs-from-default-p', but it does
not make it *look* different from `default', so doing that effectively
hides the refinements.