Emacsg: Load essential libraries independently
Multiple packages require these libraries and we don't want to
attribute the time it takes to load these libraries to whatever
dependent package happens to be loaded first.

We don't do the same for `cl-lib' because `borg' itself requires,
which is done without `use-package', i.e. without timing how long
it takes.